requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.5
freezegun==1.2.2
gunicorn==20.1.0
jamo==0.4.1
mysqlclient==2.1.1; sys_platform == 'linux'
schema==0.7.5